Refactor: Extract Examples section into a separate component
Moved the Examples section from App.js to a new Examples.jsx file for better modularity and maintainability.
